@@ -3,7 +3,6 @@ module Restrictions
     belongs_to :resource, class_name: 'Restrictions::Resource'
     belongs_to :set, class_name: 'Restrictions::Set'
 
-    validates :resource_id, presence: true, uniqueness: { scope: :set_id }
-    validates :set_id, presence: true
+    validates :resource_id, uniqueness: { scope: :set_id }
   end
 end
