@@ -3,7 +3,6 @@ module Restrictions
     belongs_to :role
     belongs_to :set, class_name: 'Restrictions::Set'
 
-    validates :role_id, presence: true, uniqueness: { scope: :set_id }
-    validates :set_id, presence: true
+    validates :role_id, uniqueness: { scope: :set_id }
   end
 end
