@@ -4,7 +4,6 @@ require File.expand_path('../boot', __FILE__)
 require 'active_record/railtie'
 require 'action_controller/railtie'
 require 'action_mailer/railtie'
-require 'active_resource/railtie'
 
 Bundler.require(*Rails.groups)
 require 'authorization'
@@ -13,11 +12,12 @@ module Dummy
   class Application < Rails::Application
     config.encoding = 'utf-8'
     config.filter_parameters += [:password]
-
+    config.eager_load = false
     config.active_support.escape_html_entities_in_json = true
-    config.active_record.whitelist_attributes = true
     config.assets.enabled = true
     config.assets.version = '1.0'
+
+    config.paths['config/database'] = File.expand_path('../../../../../../dummy_database.yml', __dir__)
   end
 end
 
